+++ test/datasift/core_api_test.rb
@@ -6,148 +6,174 @@ describe 'DataSift' do
     auth      = DataSiftExample.new
     @datasift = auth.datasift
     @data     = OpenStruct.new
-    @statuses = OpenStruct.new
-    @headers  = OpenStruct.new
 
     @data.valid_csdl    = 'interaction.content contains "test"'
     @data.invalid_csdl  = 'interaction.nonsense is not valid'
-
-    @statuses.valid        = 200
-    @statuses.bad_request  = 400
-
+    @data.invalid_hash  = 'this_is_not_a_valid_stream_hash'
   end
 
-  describe '#compile/validate' do
-    before do
-      @headers.csdl_compile = {
-          "date"              => "Thu, 30 Jan 2014 10:09:19 GMT", "content-type" => "application/json",
-          "transfer-encoding" => "chunked", "connection" => "close", "x-api-version" => "1",
-          "x-ratelimit-limit" => "10000", "x-ratelimit-remaining" => "10000", "x-ratelimit-cost" => "5"}
-
-      #valid CSDL
-      stub_request(:post, /api.datasift.com\/.*\/validate/).
-          with(:body => {"csdl" => @data.valid_csdl}).
-          to_return(:status  => @statuses.valid,
-                    :body    => fixture('validate_csdl_valid.json'),
-                    :headers => @headers.csdl_compile)
-      #invalid CSDL
-      stub_request(:post, /api.datasift.com\/.*\/validate/).
-          with(:body => {"csdl" => @data.invalid_csdl}).
-          to_return(:status  => @statuses.bad_request,
-                    :body    => fixture('validate_csdl_invalid.json'),
-                    :headers => @headers.csdl_compile)
-      #valid stream compilation
-      stub_request(:post, /api.datasift.com\/.*\/compile/).
-          with(:body => {"csdl" => @data.valid_csdl}).
-          to_return(:status  => @statuses.valid,
-                    :body    => fixture('compile_csdl_valid.json'),
-                    :headers => @headers.csdl_compile)
+  ##
+  # Client
+  #
+  describe 'client' do
+    it 'raises_InvalidConfigError_without_auth_credentials' do
+      assert_raises InvalidConfigError do
+        DataSift::Client.new({})
+      end
     end
+  end
 
-    it 'test_csdl_cant_be_nil_when_validating' do
+  ##
+  # /validate
+  #
+  describe '#validate' do
+    it 'csdl_cant_be_nil_when_validating' do
       assert_raises InvalidParamError do
         @datasift.valid?(nil)
       end
     end
 
-    it 'test_csdl_cant_be_empty_when_validating' do
+    it 'csdl_cant_be_empty_when_validating' do
       assert_raises InvalidParamError do
         @datasift.valid?('')
       end
     end
 
-    it 'test_if_user_can_get_successful_validation_as_bool' do
-      assert @datasift.valid?(@data.valid_csdl), 'Valid CSDL must return true'
-    end
-
-    it 'test_if_user_can_get_successful_validation_as_hash' do
-      validation = @datasift.valid?(@data.valid_csdl, false)
-      assert_kind_of Hash, validation, 'Valid should return a hash here'
-      assert_equal @statuses.valid, validation[:http][:status], "This request should have returned status as #{@statuses.validate_csdl}"
-    end
-
-    it 'test_failing_csdl_validation' do
-      assert_raises BadRequestError do
-        @datasift.valid?(@data.invalid_csdl)
+    it 'user_can_get_successful_validation_as_bool' do
+      VCR.use_cassette('core/validate_success_bool') do
+        assert @datasift.valid?(@data.valid_csdl), 'Valid CSDL must return true'
       end
     end
 
-    it 'test_csdl_cant_be_nil_when_compiling' do
+    it 'user_can_get_successful_validation_as_hash' do
+      VCR.use_cassette('core/validate_success_hash') do
+        validation = @datasift.valid?(@data.valid_csdl, false)
+        assert_kind_of Hash, validation,
+          "Successful validation will return a hash"
+        assert_equal STATUS.valid, validation[:http][:status],
+          "This request should have returned #{STATUS.valid} status"
+      end
+    end
+
+    it 'failing_csdl_validation' do
+      VCR.use_cassette('core/validate_invalid_hash') do
+        assert_raises BadRequestError do
+          @datasift.valid?(@data.invalid_csdl)
+        end
+      end
+    end
+  end
+
+  ##
+  # /compile
+  #
+  describe '#compile' do
+    it 'csdl_cant_be_nil_when_compiling' do
       assert_raises InvalidParamError do
         @datasift.compile(nil)
       end
     end
 
-    it 'test_csdl_cant_be_empty_when_compiling' do
+    it 'csdl_cant_be_empty_when_compiling' do
       assert_raises InvalidParamError do
         @datasift.compile('')
       end
     end
-  end
 
-  describe '#usage' do
-    describe 'with valid request' do
-      before do
-        @headers.valid_usage = {
-            "date"              => "Thu, 30 Jan 2014 10:09:19 GMT", "content-type" => "application/json",
-            "transfer-encoding" => "chunked", "connection" => "close", "x-api-version" => "1",
-            "x-ratelimit-limit" => "10000", "x-ratelimit-remaining" => "10000", "x-ratelimit-cost" => "25"}
-
-        #valid /usage request
-        stub_request(:post, /api.datasift.com\/.*\/usage/).
-            with(:body => {:period => 'hour'}).
-            to_return(status:  @statuses.valid,
-                      body:    fixture('usage_current.json'),
-                      headers: @headers.valid_usage)
-      end
-
-      it 'can get users usage' do
-        @datasift.usage
-        assert_requested(:post, "https://api.datasift.com/v1/usage", :body => {"period"=>"hour"})
+    it 'successful_compilation_returns_hash' do
+      VCR.use_cassette('core/compile_success') do
+        response = @datasift.compile @data.valid_csdl
+        assert_kind_of Hash, response,
+          "Successful compilation will return a hash"
+        assert_equal STATUS.valid, response[:http][:status],
+          "This request should have returned #{STATUS.valid} status"
       end
     end
   end
 
+  ##
+  # /usage
+  #
+  describe '#usage' do
+    it 'can_get_users_usage' do
+      VCR.use_cassette('core/usage_success') do
+        response = @datasift.usage
+        assert_equal STATUS.valid, response[:http][:status]
+        assert_kind_of Hash, response
+      end
+    end
+  end
+
+  ##
+  # /dpu
+  #
   describe '#dpu' do
     before do
-      @data.dpu_hash_valid = '145ea24a4d83a14ecb9077b831f14809'
-      @headers.dpu = {
-          "date"              => "Thu, 30 Jan 2014 10:09:19 GMT", "content-type" => "application/json",
-          "transfer-encoding" => "chunked", "connection" => "close", "x-api-version" => "1",
-          "x-ratelimit-limit" => "10000", "x-ratelimit-remaining" => "10000", "x-ratelimit-cost" => "5"}
-
-      #valid /dpu request
-      stub_request(:post, /api.datasift.com\/.*\/dpu/).
-          with(:body => {:hash => @data.dpu_hash_valid}).
-          to_return(status:  @statuses.valid,
-                    body:    fixture('dpu_valid.json'),
-                    headers: @headers.dpu)
+      VCR.use_cassette('core/before_dpu') do
+        @hash = @datasift.compile(@data.valid_csdl)[:data][:hash]
+      end
     end
 
-    it 'can get dpu cost for valid stream' do
-      @datasift.dpu(@data.dpu_hash_valid)
-      assert_requested(:post, "https://api.datasift.com/v1/dpu", body: {hash: @data.dpu_hash_valid})
+    it 'can_get_dpu_cost' do
+      VCR.use_cassette('core/dpu_get_cost') do
+        response = @datasift.dpu @hash
+        assert_equal STATUS.valid, response[:http][:status]
+      end
+    end
+
+    it 'cannot_get_dpu_cost_for_invalid_hash' do
+      VCR.use_cassette('core/dpu_throw_badrequest') do
+        assert_raises BadRequestError do
+          @datasift.dpu @data.invalid_hash
+        end
+      end
+    end
+
+    it 'requires_at_least_one_param' do
+      assert_raises ArgumentError do
+        @datasift.dpu
+      end
     end
   end
 
-  describe '#balance' do
+  describe '#dpu for Historics' do
     before do
-      @headers.balance = {
-          "date"              => "Thu, 30 Jan 2014 10:09:19 GMT", "content-type" => "application/json",
-          "transfer-encoding" => "chunked", "connection" => "close", "x-api-version" => "1",
-          "x-ratelimit-limit" => "10000", "x-ratelimit-remaining" => "10000", "x-ratelimit-cost" => "25"}
-
-      #valid /dpu request
-      stub_request(:post, /api.datasift.com\/.*\/balance/).
-          to_return(status:  @statuses.valid,
-                    body:    fixture('balance.json'),
-                    headers: @headers.balance)
+      VCR.use_cassette('core/before_historic_dpu') do
+        @hash = @datasift.compile(@data.valid_csdl)[:data][:hash]
+        @historic = @datasift.historics.prepare(
+          @hash,
+          Time.now.to_i - 10800,
+          Time.now.to_i - 7200,
+          'Ruby test suite',
+          'tumblr',
+          10
+        )
+      end
     end
 
-    it 'can get account balance' do
-      @datasift.balance
-      assert_requested(:post, "https://api.datasift.com/v1/balance")
+    after do
+      VCR.use_cassette('core/after_historic_dpu') do
+        @datasift.historics.delete @historic[:data][:id]
+      end
+    end
+
+    it 'can_get_dpu_cost_for_historic' do
+      VCR.use_cassette('core/historic_dpu') do
+        response = @datasift.dpu('', @historic[:data][:id])
+        assert_equal STATUS.valid, response[:http][:status]
+      end
     end
   end
 
+  ##
+  # /balance
+  #
+  describe '#balance' do
+    it 'can get account balance' do
+      VCR.use_cassette('core/balance_get') do
+        response = @datasift.balance
+        assert_equal STATUS.valid, response[:http][:status]
+      end
+    end
+  end
 end