+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 CHANGELOG
 ================================
+## v.3.2.0 (2015-05-13)
+### Added
+* Support for [PYLON API](http://dev.datasift.com/pylon/docs/api-endpoints)
+* Support for [Account Identities API](http://dev.datasift.com/pylon/docs/api/account-api)
+* Adds opts for Managed Sources create/update endpoints to allow passing of 'validate' param
+* Support for HTTP 409 Conflict error
+* Support for HTTP 410 Gone error
+* Comprehensive Yard Docs for all classes and methods
+* Due to API v1.1 change, we've added support for ```include_finished``` and ```all``` parameters when making calls to the ```/push/get``` API endpoint
+* Due to the API v1.1 change, a ```delivery_count``` field has been added to the main object, and individual chunks in the response from ```/historics/get``` calls
+* Due to the API v1.1 change, an ```interaction_count``` field has been added to the respomse from ```/push/get``` API calls
+* The ```dpu()``` method now also accepts the ```historics_id``` parameter, which allows you to get the DPU cost of an Historics query
+
+### Changed
+* Some refactoring for the Rubocop across the library
+* We now use [VCR](https://github.com/vcr/vcr) in our test suite for all outbound API calls
+* Updated some methods due to deprecations in the Ruby language
+* Use DataSift API v1.1 by default
+* Due to the API v1.1 change, /usage and /balance API calls will now return an empty Object rather than an empy Array when there is no data available
+
+### Deprecated
+* Support for Ruby 1.x is being dropped in the next major release; 4.0.0
+* Due to the API v1.1 change, the ```volume_info``` field has been removed from ```/historics/get``` API calls
 
 v.3.1.5 (2015-04-16)
 --------------------